spell-checker/requirements.txt
fastapi
uvicorn
pyspellchecker
